// src/store/type.ts
export enum ModalActionType {
    OPEN_ADD_FRIEND_MODAL = 'OPEN_ADD_FRIEND_MODAL',
    CLOSE_ADD_FRIEND_MODAL = 'CLOSE_ADD_FRIEND_MODAL',
    OPEN_EDIT_POST_MODAL = 'OPEN_EDIT_POST_MODAL',
    CLOSE_EDIT_POST_MODAL = 'CLOSE_EDIT_POST_MODAL',
    OPEN_USER_IMAGE_MODAL = 'OPEN_USER_IMAGE_MODAL',
    CLOSE_USER_IMAGE_MODAL = 'CLOSE_USER_IMAGE_MODAL',
    OPEN_COMMENT_MODAL = 'OPEN_COMMENT_MODAL',
    CLOSE_COMMENT_MODAL = 'CLOSE_COMMENT_MODAL'
}